requirements.txt
asgiref==3.2.3
backcall==0.1.0
botocore==1.13.44
colorama==0.4.3
decorator==4.4.1
Django==3.0.1
django-extensions==2.2.5
django-filter==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
ipdb==0.12.3
ipython==7.10.2
ipython-genutils==0.2.0
jedi==0.15.2
jmespath==0.9.4
Markdown==3.1.1
parso==0.5.2
pickleshare==0.7.5
prompt-toolkit==3.0.2
Pygments==2.5.2
pynamodb==4.2.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
